Avoid adding open walls that are occluded by the nearest wall
There are plenty of cases where a wall is completely hidden behind other open walls and therefore cannot contribute to
the result. Instead of marking these as open, we can skip them completely and avoid a lot of expense during the sweep.

In order to keep the check cheap, walls are only checked against the nearest wall, not against all open walls that might
hide them.