Add support for parameterized timer IDs
This is inspired by loggers, though based on `String.format()`. It allows using timers with dynamic names without
needing to clutter calling code with `CodeTimer.isEnabled()` checks to avoid the overhead when not needed.

The few places that used `CodeTimer.isEnabled()` for this purpose have been updated to use the parameterized form, and
`isEnabled()` is now only used to check if timings should be reported.