Consistently add light options to Campaign Properties
When iterating over the segments of a light source, instead of separate variables to track previous state we now have a
map that maintains all properties from the previous segment. In the loop we just update the map, and then have a uniform
way to output them to the `StringBuilder`.

The order now matches documentation: shape+arc/offset, owner/gm, radius+colour+lumens. Parameters are only included if
different from the previous occurrence (or from the default for the first occurrence).