Avoid JTS Geometry in vision sweep
The sweep is now based on `LineSegment` instead of `LineString` as it is a bit easer to work with and avoids some of the
cost associated with `Geometry`.

The intersection check for whether to include a blocking segments is now against the vision bounding box rather than the
precise vision geometry. There is a tradeoff here as more segments can be excluded the more precise we are, but in
practice the extra precision does not gain us nearly as much as it costs.

`VisibilityProblem` is now responsible for collecting and organizing segments. `VisionBlockingAccumulator` only decides
how to treat each topology type, but otherwise forwards to `VisibilityProblem`. This avoids some expense around building
large lists of segments.