Factor light & sight syntax out of campaign properties dialog
New `LightSyntax` defines the syntax of lights, handling both parsing and rendering as strings. The logic is essentially
what it was in `CampaignPropertiesDialog`, but has been decomposed so that lights can be used without or without
categories.

In the same vein, `SightSyntax` defines the syntax of sights, though not extra functionality has been built into it.