Better visibility graph representation
A new `EndpointSet` class is responsible for maintaining the endpoints and walls for the visibility sweep. It avoids any
use of endpoint hashing in favour of retroactively merging the few duplicated endpoints we
receive. `VisibilitySweepEndpoint` now references other `VisibilitySweepEndpoint` to define walls instead of using
`LineSegment`, which saves on some redundancy and make it really easy to validate the graph.