Add a trivial personal server instead of using a real server
All servers now implement `IMapToolServer` with any generic code using that. Similarly, connections to the server are
represented by `IMapToolConnection`, which might be a `NilMapToolConnection` that is for use with personal servers.

`ServerHandshake` has been updated to hold onto a `MapToolServer` instance rather than looking it up all the time. This
avoids any possibility of a server handshake picking up the wrong server in cases where the handshake is slow, and a
server restart happens in the meantime.

There is no more concept of a "personal server config". Instead a personal server is a trivial concept of a server that
doesn't require a socket configuration or player passwords or anything since it is not a real server that can be
connected to.