Rework PlayerDatabaseFactory
The main change here is that `PlayerDatabaseFactory` no longer carries state about the current player database, it just
provides instances of `PlayerDatabase`. The current database is maintained in `MapToolClient` instead. There are a lot
of change listeners that aren't needed anymore as a result and the whole flow of player databases is easier to follow.

Surprisingly, `PersonalServerPlayerDatabase` was essentially unused and has been removed. Although it was set when
starting a personal server, it would be promptly replaced by a `LocalPlayerDatabase` once the handshake
completed. During that short time it didn't do anything special, so we just use `LocalPlayerDatabase` now.

Some methods in `PlayerDatabase` were specific to server-side use. These now live in the `ServerSidePlayerDatabas`
interface. This simplifies `LocalPlayerDatabase` which is only used client-side.

The `Players` constructor now takes a `PlayerDatabase` instead of having to look it up for each operation.