Refactor connection creation
`MapTool.createConnection()` and `createLocalConnection()` are very similar and have been consolidated by introducing a
new `installClient()` method.

The ordering of `server.start()`, `client.start()`, and `installClient()` is now more definite. When starting a server,
the ordering is always:
1. Create and set the server
2. Create and set the client
3. Start the server
4. Start the client

Part of that is a change where we always connect immediately to a local server. This avoids any possibiliy that we
momentarily hang onto an old client after starting a new server.