Speed up FogUtil.exposeLastPath() by improving area unions
The first big win is to avoid calculating a separate union for `visionArea` and `meta`, instead calculating `visionArea`
as a union, then adding it to `meta`.

The second win is to use `GeometryUtil.destructiveUnion()` to significantly reduce the cost of creating `visionArea` for
long paths.

New timers have been added to measure how long it takes to get all the visible areas along the path, to union them up,
and to add them to the token's exposed area.