Improve endpoint sorting performance
The new comparator is based on point orientation around the origin rather than directly depending on angle. This is much
cheaper, and means we don't have to cache calculations on `VisibilitySweepEndpoint` thus making it simpler.

`EndpointSet` also buckets endpoints according to which eighth of the plane the point is in. Sorting can be done on one
bucket at a time which reduces the number of comparisons needed. The buckets also remove the need for some special cases
in the comparator, further streamlining the sort.