Fix issue with vision bounds not including the vision origin
The vision sweep is only well-behaved if it has bounds surrounding the origin. It is not enough that we use the vision
bounds since certain cases may not guarantee that the vision shape include the vision origin. So we now make sure to
explicitly expands the sweep bounds to include the origin.

Also some separation of concerns: `EndpointSet` has no reason to know anything about vision, not even just to provide
initial bounds. So now it only tracks the bounding box of its contents and nothing else.