Ensure valid geometry is returned by GeometryUtil.toJts()
Partly this is just a matter of picking a more reasonable precision. But we also now make `GeometryUtil.toJtsPolygons()`
the authoritative source for geometry, with `GeometryUtil.toJts()` merely packaging that as a `MultiPolygon`. A
defensive change to flatten any curved areas also guards against future usage that might run afoul of the current
assumption that that input is only polygonal areas.