Minimize Handshake interface
Since the clientserver code does not understand what a `Player` is, we can't have `Handshake.getPlayer()`. Since this
method is only used called on `ServerHandshake`, the method has been moved there and removed from `Handshake` and
`ClientHandshake`.

Additionally, the `Handshake` interface does not need to inherit from `MessageHandler`. Implementations can be
`MessageHandler` if they need to be (and they current do), but nothing about the handshake process inherently requires
this to be true.