Rework LineCellTemplate path and bounds
`calcPath()` has been simplified to essentially Bressenham's algorithm. It only calculates and returns a path, while
`getPath()` (previously unused) will call it if needed.

`getBounds()` now correctly accounts for the path's quadrant and the position of the first vertex to produce a correctly
located and sized bounding box. This method was also simplified to only look at the two endpoints rather than iterating
over the entire line.

All fields except `pathVertex` are now marked as `transient` since they can be recovered from only `getVertex()` and
`pathVertex`. Correspondingly, the quadrant is not longer part of the DTO, and the serialized XML is much smaller.

With the above changes in, it was natural to simplify `LineCellTemplateTool` to avoid a lot of case work, especially in
not relying on mutations of `getPathVertex()`.