Remove renderer from BlastTemplate
This renderer was still in use, but it adds some unnecessary constraints to BlastTemplate. Mainly the issue is that it
had to modify the renderer's shape, and had to override `draw()`/`drawBackground()` instead of the `paint()` method
whichi is more idiomatic for templates.

Now we don't even bother keeping a shape up-to-date as we just create a new rectangle whenever it is needed. This is not
expensive at all, especially compared to what some other templates have to do.