Give DrawnElement a copy constructor
To enable this, `Drawable` now has a `copy()` method, which is consistently implemented in child classes as delegating
to each implementation's copy constructor. The copy constructor always keeps the same ID, but callers can now modify the
ID through a new method `Drawable.setId(GUID)` for those cases where the ID needs to be given a differnt ID from the
original.