Fixup of setTableEntry performance
FullBleed complained on Discord about the performance of this MTscript
function.  It turns out that the algorithm is O(n), so putting 3200
entries in a table gets seriously slow.  Fixing this will require the
supporting `LookupTable` class to have its own `setTableEntry` method.
This commit is some simpler formatting and refactoring fixes; the next
commit will be more substantial in regards to the actual fix.