Fix tests that relied on app state not being set up properly
One test was relying on the basic campaign (`CampaignFactory.createBasicCampaign()`) being set in the app state, as it
needed properties from this basic campaign. Now that we actually set an empty campaign by default, this test broke. It
was fixed by simply setting the campaign during test setup instead of implicitly relying on it being in the expected
state already.

One test for lenient JSON parsing was attempting to test fallback behaviour for the parser. In order to trigger the
fallback, it was relying on a `NullPointerException` due to a local player not having been set up since app
initialization was not done. But now that we always have a player available through `MapToolClient`, this exception is
not thrown and the test fails. The data for the test is also not valid for the test case, and if used in real MapTool
token properties it would not have the asserted result. So this data has been replaced by data that is expected to
produce the lenient JSON result.

Another few tests relied on special logic in `MapToolVariableResolver` that allows setting token properties in tests
without going through a `ServerCommand` as that requires frame state and others to be set up. It does this by checking
whether `MapTool.getServerCommand() == null`, a case which no longer holds. As a result, `NullPointerException`s were
being thrown where they weren't before. This was fixed by changing the tests to explicitly use a variable resolver that
writes directly to tokens and never goes through `ServerCommand`. In the future this could be mitigated by having
different `ServerCommand` depending on the situation, or by reducing the amount of dependence our model code has on the
frame.