Better encapsulate server policy in MapToolClient
Callers can no longer directly modify the result of `MapToolClient.getServerPolicy()`, but must instead write it
back. The good news is we always technically did this, so that wasn't a big ask. To support this, `ServerPolicy` now has
a copy constructor.

The methods on `MapTool` for updating the server policy have been removed in favour of the methods on `MapToolClient`
and `ServerCommand`.