Bring back PersonalServerPlayerDatabase
Merge PersonalServer back into MapToolServer

Now that `MapToolServer` supports direct local connections, we have no need for a specialized (and not fully functional)
`PersonalServer` just to avoid start up costs. Instead we have a `NilServer` in `clientserver` that is the simplest
fully functional `Server` implementation that never yields new connections.

This also marks the return of `PersonalServerPlayerDatabase`, but unlike before it is actually put to use by the
personal server.