Be more careful about calling setCampaign()
It should not be called for all personal starts because the conclusion of the "handshake" will involve setting it
through a `SetCampaignMsg`. This has been fixed so that we only directly call it for the initial personal server with
the default campaign so that the user is not staring at a blank screen.