Do not require handshakes for local connections
Since the server already knows who the local player is, avoiding the handshakes makes startup times
near-instantaneous.

We also connect a server and local client through a `DirectConnection` that is much like any other `Connection` but
sends messages through `Queue<>` rather than the network stack. This avoids the need for `NilMapToolConnection` which
made this a special case at the application level.