Threading improvements for connections
The existing message queue (`addMessage()`/`nextMessage()`) was a thread-unsafe FIFO queue. It looks like the intent was
to fairly share the connection between channels, but this was not achieved. It would be a good idea, but for now we just
replaced all that with a simple `BlockingQueue<>`.

Use of the queue has been simplified. `nextMessage()` is now blocking, so threads do not have to explicit
`wait()`/`notify()` to avoid spin locking, nor do they have to check `hasMoreMessages()`. A timeout of 10ms is included
just to be extra sure that the receive thread will not block indefinitely.

Some flags were updated to be atomic where needed. And in `WebRTCConnection`, we don't worry about the peer connection
status anymore, just the data channel status.