Avoid transmitting the campaign to the local client on server start.
For large campaigns, a local client can wait quite a while for the server to serialize and send the campaign, despite
the client already having the same campaign. So now we skip that, simply invoking `setCampaign()` with the client's copy
of the campaign.

As an added perk, this avoids the client jumping to a random map when a server is started.