Clean up disconnect logic for sockets and WebRTC
For `SocketConnection`, only the receive thread calls `fireDisconnect()`, which it now does whenever it exits. Thje send
thread can call `close()` in those cases where it needs to bail.

For `WebRTCServer`, we no longer call `MapTool.disconnect()` as that implies a clean disconnection. Instead we only call
`stopServer()` which signals an unexpected disconnect.