Do not null out the server field on disconnect, just stop it.
This matters because a `null` server is presumed to indicate a remote connection. In reality, it could mean an
unexpected situation occurred for a local server, even a personal server. This changes allows a new server to be started
in these cases, avoiding some possibilities for the user to be "soft locked" after an error.

We now explicitly null out the server field when connecting to a remote server.