Stronger exception guarantees when starting servers
For both `MapToolClient.start()`, if an `IOException` is thrown we now guarantee that the state will be `Closed`. This
means the caller doesn't have to worry about cleaning up the client by calling `.close()` themselves on an object with
some indeterminate state. Similar thing for `MapToolServer.start()` - its state will be set to `Stopped` if an
`IOException` is thrown.

`MapTool.startServer()` now relies on this by catching any `IOException` from the client. In this case it stops the
server before propagating the exception. This makes sure both client and server are in a determinate state if
`startServer()` fails.

The change to `SocketServer.start()` is only to make it more clear what the state will be in case of a failure.