Reorganize logic for handshake completion and disconnection handlers
onComplete callbacks ow receive a flag indicating whether the handshake was successful. This allows us to bail early and
to avoid proceeding with logic that assumes a successful connection. On the disconnect side, some logic was moved around
to where it makes more sense, in particular expected disconnects resulting from `AppActions.DISCONNECT_FROM_SERVER`.

Clients are now created after their respective servers if any. This isn't a huge deal, but is more intuitive and allows
us a bit of an easier time in the client disconnect handler.