Base zone import on the copy constructor
The copy constructor already guarantees that everything has independent copies made - and if it missed something that
would be a bug. So it's appealing to use it for map imports where we want the same thing. Also preserve the creation
time in the copy constructor when `keepIds` so that it is preserved when campaigns are copied and such.

`Zone#import()` no longer needs to do as much. All it does not is clear the initiative list.