Clean up walker and points
`ZoneWalker.replaceLastWaypoint()` does not need to return a value, so that has been removed.

`ZoneWalker.getPath()` and `SelectionSet.getGridlessPath()` never return `null`, so mark them as such and clean up
callers to not check for it.

`AbstractPoint` is now sealed with `ZonePoint` and `CellPoint`. We do a lot of casework that already assumes only those
two possibilities, and now that is formal. It will also allow for pattern matching.