MeasureTool no longer modifies its path
Instead it keeps its own last point that it can freely modify but that is kept out of the path until committed as a
waypoint.

Gridless path rendering is also cleaned up by using a Path2D. Even better would be to delegate to
`ZoneRenderer.renderPath()` like we do for gridded paths, but that current assumes it needs to adjust for a token
footprint, which `MeasureTool` does not require.