SelectionSet no longer modifies its path
Just like `MeasureTool`, `SelectionSet` keeps its own current point that it modifies and only adds to the path when set
as a waypoint. If the final path is requested, the path is copied and the current point is added as the final waypoint.

A subtle aspect of this change is that all points in the `Path` are waypoints and there are no "dangling" cell points.