Create a new append-only API for Path
Now that all users of `Path` do not modify the `Path` and are guaranteed to not have dangling cell points, we can have a
new mutating API for Path:
- One method for appending a waypoint (implies adding a cell point too)
- Oen method for appending a partial path, the last point of which is treated as a waypoint.

This gives the following invariants for `Path`:
1. If there are no waypoints, then there are no cell points.
2. The first cell point in the path is a waypoint.
3. The last cell point in the path is a waypoint.
4. The waypoints are a subsequence of the cell points.

Before this change, there we no invariants for `Path` - the cell points and waypoints could vary arbitrarily depending
on the caller. The downside we still have is that the representation does not reflect this invariant. E.g., there is no
way accurately pull the path back apart into its partial paths.