Refactor how we apply moves to follower tokens
First of all, we copy the leader token before making any changes to the movement set. Otherwise we run the risk of
modifying the leader partway and getting inconsistent results depending on iteration order.

Next we pull the responsibility for derive the path out of `Token`. Previously, `Token.applyMove()` had to be aware of a
bunch of information that is really specific to the renderer state, only to pass it along to `Path.derive()`. We now
instead just make the renderer derive the path, then set it on the token. This also removes the need for
`Token.applyMove()`.

Finally, we push the calculation of the "cell offset" into `Path.derive()`. There is casework in the renderer and in
`Path.derive()` that has to line up or else things don't work. So it is simpler to just have `Path.derive()` do the
calculation in those cases where it is needed.