Rework path derivation
The calculation for how to derive a follower path is quite different now (and simpler). When both tokens are STG or both
are not STG, the calculation is the same: simply apply the offset to all path points. This lets the follower's path have
the exact same shape as the leader.

When a non-STG token is following a STG leader, we reflect the waypoints of the leader in the follower path (offset, of
course). This does not keep the full pathfinding shape when pathfinding around terrain, but at least the path is not
arbitrarily chosen and does not suffer issues with zig-zagging.

When a STG token is following a STG leader, we again reflect the waypoints. But in this case we have to snap each
resulting point to the grid, and need to fill in partial paths between each waypoint. As before, this pathfinding is
simple, i.e., it is not an A*-based algorithm.