Fix edge projection for stretched hex grids
The edge projection was not always recalculated when the dimensions change. On top of that, it is not always half the
edge length - that's only for regular hexagons. Now we always recalculate edge projection and edge length together, and
relate them to the diameter propertly.

There was one place where edge length was used where the diameter should have been used.

Also remove a bunch of unused and commented out code, and caches for things that don't need to be cached.