Fix splash screen transparency on Linux
The problem is that while `JFrame` supports per-pixel transparency, and `JFXPanel` also supports transparency,
`JFXPanel` refuses to be transparent if used in a transparent `JFrame`. But we don't need a `JFXPanel` here since we're
just displaying an image, so it has been replaced with a `JPanel` that does exactly that and no more.

It was also necessary to be a bit more particular in how `SplashScreen` is initialized. We must get the screen
dimensions _after_ `setUndecorated()` and `setType()`, and must call `setLocationRelativeTo(null)` after setting the
windw bounds. The reasons are not entirely clear, but may be related to weaknesses in window transparency for
multi-monitor setups on X.

`SplashScreen` is also not responsible for setting its own visibility. The caller decides when to do that now. We can
also greatly limit the portion of `SplashScreen` that needs to run on the JavaFX thread now that we only use JavaFX for
generating the image.