Add js getRawProperty and getEvaluatedProperty
Previously you could tell that the property was unset by if it returned
the string "null", so getRawProperty gives that feature back
but returning a null so you can use ?? to default it yourself.

getEvaluatedProperty adds the ability to handle calculated defaults.
getProperty was not made to use getEvaluatedProperty
because a javascript function that knows the campaign doesn't use
calculated defaults could use it to save on performance,
but also that switching to evaluating by default is more likely to be a
breaking change.

The code for handling a missing parameter in earlier versions is awkward
so it's likely that code that uses getProperty is used in campaigns
where there is no default value to be reset to and all properties are
explicitly set so doesn't trip the awkward behaviour.

However, if this code contained objects that could be mis-parsed as
macros beginning to use macros by default would be a breaking change.