Add new Grid methods for snapped facing
Tools no longer call `Grid.getFacingAngles()`, which requires each caller to interpret the resposne to get the nearest
or next facing. Now there are two new methods in on `Grid`:
- `int nearestFacing(int facing, boolean faceEdges, boolean faceVertices)`: snaps a facing to one of the standard
  facings.
- `int nextFacing(int facing, boolean faceEdges, boolean faceVertices, boolean clockwise)`: snaps a facing as in
  `nearestFacing()`, but also advances one step clockwise or counterclockwise.

There is no new behaviour introduced by this change.

Also included are tests cases for interesting input facings. These make it clear there are some bugs in the current
behaviour, such as -179° being considered closer to -135° than to 180° (these are 44° and 1° away, respectively).