Reimplement nearestFacing() and nextFacing() mathematically
Instead of a provide a hardcoded array of possible facings, we now calculate the nearest facing using some modular
arithmetic and rounding. This fixes the bugs mentioned in the prior commit, and test cases have been updated to match.

Internally to `Grid`, the use of `getFacingAngles(boolean faceEdges, boolean faceVertices)` and been replaced with
`snapFacingInternal(int facing, boolean faceEdges, boolean faceVertices, int addedSteps)`, which is enough to provide
the functionality of both `getNearestFacing()` and `getNextFacing()`.