Fix hex light construction
The proportions and center were just all wrong. `Grid.getShapedArea()` should always treat (0, 0) as its origin for the
light, so no adjustments are needed to account for the grid or token shape.

Also fixed a bug where the hex radius would measure to the vertices instead of the edges (compare with square lights
that measure to the edges).

Since we're modifying the hex code anyways, also refactored to be more direct. We don't need trig - there's only six
points and they're always the same up to scale. Well, until the day we account for stretched hex grids.