Convert all AppPreference preferences to new type
All preferences are now instances of `Preference<T>`. The few types that had additional update logic not use the change
listener to notify observers about it.

In order to really make this change effective, we no longer store any colors component-wise since they can easily be
stored as a single integer. For those preferences that once had component-wise storage, the static initializer ensures
backwards compatibility by detecting whether a new value is present and loading the old values if not.