Add support for instance-based tools
Existing tools are all class-based, i.e., are identified by a class, and state is tracked in terms of classes. Now it is
possible to define tools by passing instances. The only difference is that instance-based tools cannot be looked up by
their class, but that functionality is only used for a select few transitions.