Make SocketConnection take a supplier function
SocketConnection has two constructors, either one wrapping an existing
socket or one that creates the socket when started.

Creating the socket from just the host and port is insufficiently
flexible for supporting SSL sockets and since the only difference for
how to handle SSL sockets is how they are created, making creation more
flexible is the minimal change required to support them.