Improve paint handling in GdxRenderer
The `GdxPaint` record is now a top-level type not embedded in `ZoneCache`. `ZoneCache` is also no longer responsible for
converting `DrawablePaint` to `GdxPaint` as that has nothing to do with per-zone caching. Instead, it now provides
`#getPaintTexture(MD5Key)` to do a cached lookup of a paint texture. Meanwhile, `GdxRenderer` adds
`#getPaint(DrawablePaint)` to convert `DrawablePaint` to `GdxPaint` based on `ZoneCache#getPaint(MD5Key)`.

This also include a bugfix to `GdxRenderer#renderFog(PlayerView)` to not modify the color of the fog paint. Although the
`GdxPaint` objects are temporary in nature, the object in `GdxPaint#color()` may be a global object and thus should not
be modified. In particular, `ZoneCache#getPaint(DrawablePaint)` used to set this field to `Color.WHITE` for texture
paints, resulting in drastic consequences whenever textured fog was rendered. As a safety measure, we also no longer
directly set this field to `Color.WHITE`, but always create a temporary color.

`GdxRenderer#fillViewportWith(GdxPaint)` is now `#fillViewportWith(Color, TextureRegion)` so we can freely modify colors
coming from `GdxPaint` without having to wrap it in a new `GdxPaint`.