Change GDX paints to be straight Texture instead of TextureRegion
A `Texture` has the ability to repeat, which is vital for paints. A `TextureRegion` can do this if it represents a whole
texture, but in general is not suitable as it only represents a portion of a texture with no way to repeat without
including other regions.

In `AreaRenderer`, inject the white pixel texture directly instead of assuming the `ShapeDrawer` has it set as its
current region,